% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_gpsm.R
\name{dr_gpsm}
\alias{dr_gpsm}
\title{Doubly-robust GPS matching estimator}
\usage{
dr_gpsm(
  data,
  treatment,
  treatment_ref = NULL,
  covariate,
  outcome,
  gps_model = c("logit", "rf", "gbm", "gam"),
  outcome_model = c("none", "lm", "rf"),
  folds = 2,
  nboot = 500,
  hist = FALSE,
  hist_path = NULL,
  gps_params = NULL
)
}
\arguments{
\item{data}{Data frame with treatment, covariates, and outcome variables}

\item{treatment}{Column index of the treatment variable}

\item{treatment_ref}{Reference treatment level (default = first level)}

\item{covariate}{Numeric vector of covariate column indices}

\item{outcome}{Column index of the outcome variable}

\item{gps_model}{Choice of GPS model: 'logit', 'rf', 'gbm', or 'gam'}

\item{outcome_model}{Choice of outcome model for doubly-robust adjustment: 'none', 'lm', or 'rf'}

\item{folds}{Number of folds for cross-fitting (default = 2)}

\item{nboot}{Number of bootstrap replications (default = 500)}

\item{hist}{Logical; TRUE to save GPS histogram}

\item{hist_path}{Character string specifying file path (pdf, png, jpg) to save the histogram (used if hist = TRUE)}

\item{gps_params}{Optional named list of hyperparameters per GPS model.
For example: list(
rf  = list(ntree=1000, mtry=3, sampsize=..., nodesize=2, maxnodes=..., nPerm=5),
gbm = list(n.trees=4000, interaction.depth=4, shrinkage=0.01,
cv.folds=5, n.minobsinnode=5, n.cores=2),
logit = list(maxit=200, decay=1e-3, trace=FALSE),
gam   = list(df_max=6, maxit=80, trace=FALSE))}
}
\value{
A list with components:
\describe{
\item{estimate}{Named numeric vector of doubly-robust ATE estimates for each contrast}
\item{ci_lower}{Named numeric vector of lower 95\% confidence bounds}
\item{ci_upper}{Named numeric vector of upper 95\% confidence bounds}
}
}
\description{
Doubly-robust GPS matching estimator
}
