% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_pre_process.R
\name{gps_pre_process}
\alias{gps_pre_process}
\title{Generalized Propensity Score Model fitting}
\usage{
gps_pre_process(
  data,
  treatment,
  treatment_ref = NULL,
  covariate,
  gps_model = c("logit", "gbm", "gam"),
  gps_params = NULL
)
}
\arguments{
\item{data}{Data frame containing treatment, covariates, and outcome variables}

\item{treatment}{Column index of the treatment variable}

\item{treatment_ref}{Reference treatment level (optional; default is the first level)}

\item{covariate}{Numeric vector of covariate column indices}

\item{gps_model}{Choice of GPS model: 'logit', 'rf', 'gbm', or 'gam'}

\item{gps_params}{Optional named list of hyperparameters per GPS model.
For example: list(
logit = list(maxit=200, decay=1e-3, trace=FALSE),
gam   = list(df_max=6, maxit=80, trace=FALSE),
gbm = list(n.trees=4000, interaction.depth=4, shrinkage=0.01,
cv.folds=5, n.minobsinnode=5, n.cores=2))}
}
\value{
A data frame containing the original data plus estimated GPS columns (gps_) and log-transformed GPS columns (loggps_)
}
\description{
Generalized Propensity Score Model fitting
}
